chore: use classes for page objects
