--- suite-native/app/e2e/pageObjects/myAssets.ts
@@ -1,8 +1,0 @@
-export const tapAddAccountButton = async () => {
-    await element(by.id('@screen/mainScrollView')).scrollTo('top');
-    await element(by.id('@myAssets/addAccountButton')).tap();
-};
-
-export const navigateToMyAssets = async () => {
-    await element(by.id('@tabBar/AccountsStack')).tap();
-};
